aduana_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.58.1
httpx==0.27.2
Pillow==10.1.0
python-multipart==0.0.6
